# Change log level in test environment to speed up tests