--- a/config/environments/test.rb
+++ b/config/environments/test.rb
@@ -77,6 +77,9 @@ Rails.application.configure do
 
   # Raise error when a before_action's only/except options reference missing actions
   config.action_controller.raise_on_missing_callback_actions = true
+
+  # disable logging in tests, for speed increases. Set to :info to bring back logging
+  config.log_level = :warn
 end
 
 Paperclip::Attachment.default_options[:path] = Rails.root.join('spec', 'test_files', ':class', ':id_partition', ':style.:extension')