@josh-works 2024-04-18
@renchap I'm not sure what the least obtrusive way is for me to get a workflow approval to run tests. This PR is still in draft mode, and the change I've made so far would make tests at best perhaps 6% faster. It would get lost in the noise.

But I hope to make larger gains soon, and will certainly want to inspect the workflow output then.