@mjankowski 2024-04-19
Yup - I've done some work both on coverage and speed in the suite, but more is still needed.

I skimmed your notes and I'm curious how you are running the suite to get that ~20 min run time? Obviously it depends on exact hardware and so on, but that feels high. I also see ~5 min full suite runs locally. CI is quite variable from run to run because we don't get the same underlying resources in every run.