@josh-works 2024-04-19
> I skimmed your notes and I'm curious how you are running the suite to get that ~20 min run time? Obviously it depends on exact hardware and so on, but that feels high.

@mjankowski thanks for checking the correctness of those times. When I was seeing 20 minutes for 5500 examples, something felt off, and now I have a direction to go. 

I'm running tests natively on macOS (not using docker) via:

```
$ RAILS_ENV=test bundle exec rspec spec
```

I've got 8gb of RAM that I'm throwing at tests, too. Not a ton, but that alone doesn't seem to explain it. 

I ran `bin/dev` before the tests, most stuff looks fine, just _extremely slow_. 🤔 

Could you like me to some of the spec/performance work? I peeped [your recent PRs](https://github.com/mastodon/mastodon/pulls?q=is%3Apr+sort%3Aupdated-desc+author%3Amjankowski++spec), and there's a TON going on, it was a lot to look through. I found [CI parrallelizing](https://github.com/mastodon/mastodon/pull/27663), which is great.

It's helpful to see your experience with all this. Thanks! 

I'll let you know what I learn when I figure out why my tests are running so slow.l