@mjankowski 2024-04-19
> I'll let you know what I learn when I figure out why my tests are running so slow.

Please do ... at minimum, if we can find something that's off we can either fix it if relevant, or update docs to make future people less likely to run into whatever the issue is.

I think that if you a) set `RAILS_ENV=test`, b) run something like `bin/rails assets:precompile` to get the assets in place - then running `bin/rspec` should be somewhere in the mid/high single minutes (again, contingent on resources). I'm also running directly (no docker) on a new-ish mac (M1 Max, 64GB ram) and it takes ~5 mins for full spec suite.

I think there are definitely areas where we could further reduce the number of factories being created, and probably areas where we are using factory records for specs which could just be using un-saved model instances or POROs instead. The test-prof and FPROF tooling in this area is helpful and has enabled some progress, but there's more to be done.