@mjankowski 2024-04-19
Also, more relevant to this PR ... I think the change makes sense, but might wrap it in an env var check (similar to production) so that it's easier to opt back in to logs in test env w/out making code change.

It'd also be good to get updated before/after benchmarks once you get your test env sorted.