@josh-works 2024-05-27
@mjankowski I've got it 'benchmarked' relative to my own machine (not a particularly fast macbook, but it doesn't really matter because the whole suite is only run via CI)

I'm still nudging this along locally, been exploring stackprof output, but don't have any big speed wins (40-80% faster) yet)