@mjankowski 2024-05-27
> @mjankowski I've got it 'benchmarked' relative to my own machine (not a particularly fast macbook, but it doesn't really matter because the whole suite is only run via CI)

Sounds good ... I think any benchmark is useful as long as the changes -vs- no-changes are running the same thing ... but it also seemed wise to get your env calibrated at least close to correct and that 20min time felt high.

I just did a local benchmark running a subset of the spec suite ... saw ~65s to ~70s for a few runs of main w/out the logging disable, and a few runs of ~58s to ~62s with it disabled. So, I think it's beneficial.

> I'm still nudging this along locally, been exploring stackprof output, but don't have any big speed wins (40-80% faster) yet)

Even something that's like ~2-5% faster (as long as there aren't bad trade-offs) is worth. All the small stuff adds up.

I think the parallelism (open flatware PR) and continuing to chip away at fabricator/db stuff is probably most promising.