@mjankowski 2024-11-26
Unrelated to this logging thing, on spec suite speed itself, running locally on 16-core M4 Max, I see about ~3 mins for normal full suite (via `bin/rspec` w/ no options); and about ~25s when running 16 parallel via flatware. Both of those are still 2-3x more than I'd like, but its fast enough that you can at least sort-of get in a test feedback loop with the parallel runner.

On CI, the asset compilation step, and/or the custom image setup are probably best areas for more improvement.