@ClearlyClaire 2024-12-11
I ran tests locally a couple time, with standard rspec and with flatware, both without the change and with the change.

Without the changes, and without using flatware, tests take approximately 3min44s to run. Without flatware and with the change, they took 3 minutes and 40 seconds.

With flatware, the results are proportionally a lot more variable, so I couldn't say for sure (27-31s before the change, 25-30s after).

The results are pretty unconclusive in my opinion, and I think we'd better err on the side of having available logs to inspect CI failures.