@josh-works 2025-01-29
I totally agree with the change not justifying the merge - I've got another PR inbound soon that will contain a lot more improvements, finishing the effort that started (and sorta fizzled out) in this PR. I appreciate the consideration.